ts7100: tsmicroctl: Add WDT feeder to waitcharge loop
U-Boot cannot support I2C WDT due to the fact that during ANY I2C
transaction, the driver will attempt to do a feed after every byte.
This is because I2C is slow and it makes sense. However the TS-7100
WDT will start out as armed for a long time, and a waitcharge might
cause the WDT to fire. A workaround for that is to have waitcharge
do a feed every loop.